refactor(models): split models/application.py into config/forms/submissions subpackage
